refactor: applied PR comments to modify groups and group-members

+++ src/modules/groups/groups.service.ts
@@ -1,11 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { GroupsRepository } from './groups.repository';
 import { InsertGroupDto, UpdateGroupDto } from './dtos/groups.dto';
-import { UsersRepository } from '../users/users.repository';
 
 @Injectable()
 export class GroupsService {
-  constructor(private readonly groupsRepository: GroupsRepository) {}
+  constructor(private readonly groupsRepository: GroupsRepository) { }
 
   async createGroup(data: InsertGroupDto) {
     return this.groupsRepository.createGroup(data);
@@ -23,19 +22,33 @@ export class GroupsService {
     return group[0];
   }
 
-  async updateGroupById(groupId: string, data: UpdateGroupDto) {
-    const updateGroup = await this.groupsRepository.updateGroup(groupId, data);
-    if (!updateGroup.length) {
+  async updateGroupById(groupId: string, userId: string, data: UpdateGroupDto) {
+    const group = await this.groupsRepository.getGroupById(groupId);
+
+    if (!group || !group.length) {
       throw new NotFoundException(`Group with ID ${groupId} not found.`);
     }
+
+    if (group[0].ownerId !== userId) {
+      throw new ForbiddenException('Only the group owner can update this group.');
+    }
+
+    const updateGroup = await this.groupsRepository.updateGroup(groupId, data);
     return updateGroup[0];
   }
 
-  async deleteGroup(groupId: string) {
-    const deletedGroup = await this.groupsRepository.deleteGroup(groupId);
-    if (!deletedGroup.length) {
+  async deleteGroup(groupId: string, userId: string) {
+    const group = await this.groupsRepository.getGroupById(groupId);
+
+    if (!group || !group.length) {
       throw new NotFoundException(`Group with ID ${groupId} not found.`);
     }
+
+    if (group[0].ownerId !== userId) {
+      throw new ForbiddenException('Only the group owner can delete this group.');
+    }
+
+    const deletedGroup = await this.groupsRepository.deleteGroup(groupId);
     return { message: 'Group deleted successfully' };
   }
 }
